// web/src/lib/api/generated/models/internal-user-create-user-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { POSKasirInternalUserRepositoryUserRole } from './poskasir-internal-user-repository-user-role';

export interface InternalUserCreateUserRequest {
    'email': string;
    'is_active'?: boolean;
    'password': string;
    'role': POSKasirInternalUserRepositoryUserRole;
    'username': string;
}